+++ src/scipp/core/groupby.py
@@ -5,11 +5,11 @@
 from typing import Optional, Union
 
 from .._scipp import core as _cpp
-from ._cpp_wrapper_util import call_func as _call_cpp_func
 
 
 def groupby(
     data: Union[_cpp.DataArray, _cpp.Dataset],
+    /,
     group: Union[_cpp.Variable, str],
     *,
     bins: Optional[_cpp.Variable] = None
@@ -23,6 +23,6 @@ def groupby(
     :return: GroupBy helper object.
     """
     if bins is None:
-        return _call_cpp_func(_cpp.groupby, data, group)
+        return _cpp.groupby(data, group)
     else:
-        return _call_cpp_func(_cpp.groupby, data, group, bins)
+        return _cpp.groupby(data, group, bins)